Moved ckeditor5 specific directory function to new module. Updated modules depending on those functions. Updated tests.
